refactor: Moving source files to `src` folder.
