+++ .gitignore
@@ -139,3 +139,4 @@ dist
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+/.swc
