--- app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,0 @@
-import { handlers } from "@/auth" // Referring to the auth.ts we just created
-export const { GET, POST } = handlers
