+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,2 @@
+import { handlers } from "@/src/auth" // Referring to the auth.ts we just created
+export const { GET, POST } = handlers
