+++ src/generatePrivateRoutes.ts
