+++ tsconfig.json
@@ -28,7 +28,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "app/lib/placeholder-data.js",
+    "src/app/lib/placeholder-data.js",
     "scripts/seed.js"
   ],
   "exclude": ["node_modules"],
